RAG/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.4.2 